Create classes

+++ src/Programming/Model/Classes/Validator.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Programming.Model.Classes
 {
@@ -31,6 +27,7 @@ namespace Programming.Model.Classes
             }
             return true;
         }
+
         public static void AssertOnPositiveValue(int value, string propertyName)
         {
             if (value < 0)
@@ -39,6 +36,7 @@ namespace Programming.Model.Classes
                     $"{propertyName} must be greater than 0");
             }
         }
+
         public static void AssertValueInRange(int value, int min, int max, string propertyName)
         {
             if (value < min || value > max)
@@ -56,7 +54,8 @@ namespace Programming.Model.Classes
                     $"{propertyName} must be greater than 0");
             }
         }
-        public static void AssertValueInRange(double value, int min, int max, string propertyName)
+
+        public static void AssertValueInRange(double value, double min, double max, string propertyName)
         {
             if (value < min || value > max)
             {
